AttnGAN-main/requirements.txt
certifi==2020.12.5
click==7.1.2
cycler==0.10.0
dataclasses==0.7
decorator==4.4.2
easydict==1.9
imageio==2.9.0
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.3
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx==2.5
nltk==3.5
numpy @ file:///C:/ci/numpy_and_numpy_base_1603480701039/work
olefile==0.46
pandas==1.1.5
Pillow @ file:///C:/ci/pillow_1603822370986/work
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.3.1
regex==2020.11.13
scikit-image==0.17.2
scipy==1.5.4
six @ file:///C:/ci/six_1605187303045/work
tifffile==2020.9.3
torch==1.7.1
torchaudio==0.7.2
torchfile==0.1.0
torchvision==0.8.2
tqdm==4.55.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
wincertstore==0.2
